Latestest.
Not likely to be many revisions left.  As I've mentioned this whole case is just for a prototype project so it's...just a testing ground at this point.

I'm going to switch most of the screw holes to threaded inserts, assuming that's not a righteous PITA.

I'd LIKE to do something about that micro_usb_cutout, which is "stupid but works."  But...feh.